polish for compatibility with Octave 5.1
had to do my own word wrapping. i enhanced the psychtoolbox routine (that i originally wrote) to accept a cell array of strings. for now it’s included. it’d be good to replace the old routine with my enhanced version. MATLAB and Octave handle dot indexing of graphic handles similarly but inconsistently. MATLAB requires capitalizaiton. Octave does not allow capitalization. So here we just use set and get, which is consistent between MATLAB and Octave.